clean-up class hierarchy and middleware constructors
