+++ studip_fuse/studipfs/path/studip_path.py
@@ -52,13 +52,6 @@ class StudIPPath(VirtualPath):
 
         super().validate()
 
-    @classmethod
-    def with_middleware(cls, list_contents_annotation, open_file_annotation, name="GenericMiddlewareStudIPPath"):
-        return type(name, (cls,), {
-            "list_contents": list_contents_annotation(cls.list_contents),
-            "open_file": open_file_annotation(cls.open_file),
-        })
-
     # FS-API  ##########################################################################################################
 
     @async_generator
